Core: Move RequestBuilder into req.go
