+++ internal/core/req.go
@@ -7,6 +7,7 @@ import (
 	"net/url"
 
 	"github.com/carlmjohnson/requests/internal/minitrue"
+	"github.com/carlmjohnson/requests/internal/slicex"
 )
 
 // NopCloser is like io.NopCloser(),
@@ -24,6 +25,39 @@ func RC(r io.Reader) NopCloser {
 func (NopCloser) Close() error { return nil }
 
 var _ io.ReadCloser = NopCloser{}
+
+type BodyGetter = func() (io.ReadCloser, error)
+
+type RequestBuilder struct {
+	headers []multimap
+	cookies []kvpair
+	getBody BodyGetter
+	method  string
+}
+
+func (rb *RequestBuilder) Header(key string, values ...string) {
+	rb.headers = append(rb.headers, multimap{key, values})
+}
+
+func (rb *RequestBuilder) Cookie(name, value string) {
+	rb.cookies = append(rb.cookies, kvpair{name, value})
+}
+
+func (rb *RequestBuilder) Method(method string) {
+	rb.method = method
+}
+
+func (rb *RequestBuilder) Body(src BodyGetter) {
+	rb.getBody = src
+}
+
+// Clone creates a new Builder suitable for independent mutation.
+func (rb *RequestBuilder) Clone() *RequestBuilder {
+	rb2 := *rb
+	slicex.Clip(&rb2.headers)
+	slicex.Clip(&rb2.cookies)
+	return &rb2
+}
 
 // Request builds a new http.Request with its context set.
 func (rb *RequestBuilder) Request(ctx context.Context, u *url.URL) (req *http.Request, err error) {
