+++ internal/core/url.go
@@ -3,7 +3,6 @@
 package core
 
 import (
-	"io"
 	"net/url"
 
 	"github.com/carlmjohnson/requests/internal/minitrue"
@@ -77,37 +76,4 @@ func (ub *URLBuilder) URL() (u *url.URL, err error) {
 		return new(url.URL), err
 	}
 	return u, nil
-}
-
-type BodyGetter = func() (io.ReadCloser, error)
-
-type RequestBuilder struct {
-	headers []multimap
-	cookies []kvpair
-	getBody BodyGetter
-	method  string
-}
-
-func (rb *RequestBuilder) Header(key string, values ...string) {
-	rb.headers = append(rb.headers, multimap{key, values})
-}
-
-func (rb *RequestBuilder) Cookie(name, value string) {
-	rb.cookies = append(rb.cookies, kvpair{name, value})
-}
-
-func (rb *RequestBuilder) Method(method string) {
-	rb.method = method
-}
-
-func (rb *RequestBuilder) Body(src BodyGetter) {
-	rb.getBody = src
-}
-
-// Clone creates a new Builder suitable for independent mutation.
-func (rb *RequestBuilder) Clone() *RequestBuilder {
-	rb2 := *rb
-	slicex.Clip(&rb2.headers)
-	slicex.Clip(&rb2.cookies)
-	return &rb2
 }
